update insert data return type

+++ src/core/include/db.h
@@ -42,8 +42,8 @@ extern "C"
     EXPORT TinyVecConnection *create_tiny_vec_connection(const char *file_path, const uint32_t dimensions);
     EXPORT IndexFileStats get_index_stats(const char *file_path);
     EXPORT VecResult *get_top_k(const char *file_path, const float *query_vec, const int top_k);
-    EXPORT size_t insert_data(const char *file_path, float **vectors, char **metadatas, size_t *metadata_lengths,
-                              const size_t vec_count, const uint32_t dimensions);
+    EXPORT int insert_data(const char *file_path, float **vectors, char **metadatas, size_t *metadata_lengths,
+                           const size_t vec_count, const uint32_t dimensions);
 
     // Internal functions
     TinyVecConnection *get_tinyvec_connection(const char *file_path);
